Updated bulk scan to use newer montoya features, updated capabilities check to use EA build number. Moved OrganiseVectors to Burp/ to account for package private methods in BulkScan
