+++ build.gradle
@@ -7,11 +7,10 @@ repositories {
 }
 
 dependencies {
-    compileOnly files("libs/montoya-api-2025.3.6.jar")
+    compileOnly 'net.portswigger.burp.extender:burp-extender-api:2.3'
     implementation 'org.json:json:20231013'
     implementation files("libs/bulkScan-all.jar")
-    compileOnly 'net.portswigger.burp.extender:burp-extender-api:2.3'
-    compileOnly 'net.portswigger.burp.extensions:montoya-api:2025.2'
+    compileOnly 'net.portswigger.burp.extensions:montoya-api:2025.3'
 }
 
 jar {
