+++ src/main/java/burp/shadow/repeater/Burp.java
@@ -4,7 +4,7 @@ import burp.api.montoya.core.Version;
 
 public class Burp {
     public enum Capability {
-        REGISTER_HOTKEY(20250200000036948L);
+        REGISTER_HOTKEY(20250300000037651L);
 
         private final long minimumSupportedBuildNumber;
